refactor: Move CLI setup helpers to Application to satisfy Onion guard
Presentation was reaching into the Domain-level CliVersionComparer
through the internal CliSetupCompatibility and CliSetupLabelFormatter
helpers, which OnionAssemblyDependencyTests.EditorUiFiles_WhenLoaded_
DoNotReferenceFacadeInternalsDirectly correctly flags as a boundary
violation: EditorUI code should reach compatibility/label workflows
through Application, not compute them directly against Domain types.

Move both helpers under Packages/src/Editor/Application/ so the
CliVersionComparer dependency now lives in an inner layer. Because they
cross an assembly boundary they become `public`, and the associated
`CliSetupCompatibilityState` struct is promoted with them. Existing call
sites in UnityCliLoopSettingsWindow, SetupWizardWindow, and
CliSetupSection already `using` the Application namespace, so no
signature or call-site changes are needed. Update
CliSetupCompatibilityTests to import from the Application namespace.

Verification:
- dist/darwin-arm64/uloop compile: 0 errors, 0 warnings
- Filtered run of StaticFacadeStateGuard, OnionAssemblyDependency, and
  CliSetupCompatibility suites: 102/102 pass
