+++ Assets/Tests/Editor/CliSetupCompatibilityTests.cs
@@ -1,6 +1,6 @@
 using NUnit.Framework;
 
-using io.github.hatayama.UnityCliLoop.Presentation;
+using io.github.hatayama.UnityCliLoop.Application;
 
 namespace io.github.hatayama.UnityCliLoop.Tests.Editor
 {
